Orquestração Smart/smartlab-task-manager/requirements.txt
django
django-jazzmin
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
python-dotenv
drf-yasg
gunicorn
